[Chore] Refactor code, Group components in components folder, Group services in services folder
