+++ src/app/app.module.ts
@@ -5,29 +5,26 @@ import { RouterModule } from '@angular/router'
 import { HttpModule } from '@angular/http'
 import { FormsModule } from '@angular/forms'
 
-import { AppRoutingModule } from './router/app-routing.module'
+import { AppRoutingModule } from './routes/app-routing.module'
 
 import { AppComponent } from './app.component';
-import { AuthComponent } from './auth/auth.component';
-import { LoginComponent } from './auth/login/login.component';
-import { RegisterComponent } from './auth/register/register.component';
-import { BucketlistComponent } from './bucketlists/bucketlists.component';
-import { BucketlistsComponent } from './bucketlists/bucketlist/bucketlist.component'
-import { BucketlistItemsComponent } from './bucketlists/bucketlist-items/bucketlist-items.component';
-import { BucketlistsService } from './bucketlists/services/bucketlists.service';
-import { BucketlistItemsService } from './bucketlists/services/bucketlist-items.service';
-import { AuthService } from './auth/services/auth.service';
-import { GlobalService } from './bucketlists/services/global.services';
+import { LoginComponent } from './components/auth/login/login.component';
+import { RegisterComponent } from './components/auth/register/register.component';
+import { BucketlistComponent } from './components/bucketlists/bucketlists.component';
+import { BucketlistItemsComponent } from './components/bucketlists/bucketlist-items/bucketlist-items.component';
+
+import { BucketlistsService } from './services/bucketlists.service';
+import { BucketlistItemsService } from './services/bucketlist-items.service';
+import { AuthService } from './services/auth.service';
+import { GlobalService } from './services/global.services';
 
 
 @NgModule({
   declarations: [
     AppComponent,
-    AuthComponent,
     LoginComponent,
     RegisterComponent,
     BucketlistComponent,
-    BucketlistsComponent,
     BucketlistItemsComponent
   ],
   imports: [
