--- src/app/auth/auth.component.html
@@ -1,16 +1,0 @@
-<div>
-
-    <div class="auth-title-bar">
-        This is the auth header.
-    </div>
-
-    <div class="auth-form-router-holder">
-        <app-login></app-login>
-        <app-register></app-register>
-    </div>
-
-    <div class="auth-footer">
-        This is the auth footer.
-    </div>
-
-</div>
