--- src/app/auth/services/auth.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { AuthService } from './auth.service';
-
-describe('AuthService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [AuthService]
-    });
-  });
-
-  it('should be created', inject([AuthService], (service: AuthService) => {
-    expect(service).toBeTruthy();
-  }));
-});
