+++ src/app/components/auth/login/login.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 
-import { AuthService } from './../services/auth.service';
+import { AuthService } from './../../../services/auth.service';
 
 export class User{
     email: string;
