+++ src/app/components/auth/register/register.component.spec.ts
