+++ src/app/components/auth/register/register.component.ts
@@ -1,6 +1,7 @@
-import { AuthService } from './../services/auth.service';
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+
+import { AuthService } from './../../../services/auth.service';
 
 
 @Component({
